# service/docs/swagger.yaml
basePath: /
definitions:
...
  dto.Decision:
    properties:
      case_id:
        type: string
      fine_decision:
        type: boolean
    type: object
...
  response.Body:
    properties:
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API Server for Traffic Police Application
  title: Traffic Police API
  version: "1.0"
paths:
...
  /expert/{id}/img:
    get:
      consumes:
      - multipart/form-data
      description: Получение фотографии эксперта по его id. Воспользоваться могут
        эксперт или директор
      operationId: expert-image-get
      parameters:
      - description: id эксперта
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Body'
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - ApiKeyAuth: []
      summary: Получение фотографии эксперта
      tags:
      - expert
    post:
      consumes:
      - multipart/form-data
      description: Добавление фотографии к профилю эксперта. Может воспользоваться
        директор или эксперт
      operationId: expert-image-upload
      parameters:
      - description: id эксперта
        in: query
        name: id
        required: true
        type: string
      - description: Фотография эксперта
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Body'
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - ApiKeyAuth: []
      summary: Добавление фотографии к профилю эксперта
      tags:
      - expert
  /expert/decision:
    post:
      consumes:
      - application/json
      description: Установка оценки случая экспертом. Воспользоваться может только
        эксперт
      operationId: expert-set-decision
      parameters:
      - description: id случая и решение эксперта
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.Decision'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Body'
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - ApiKeyAuth: []
      summary: Оценка случая экспертом
      tags:
      - expert
  /expert/get_case:
    get:
      description: Получение случая для оценки экспертом. Воспользоваться могут эксперт
        или директор
      operationId: expert-get-case
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: file
        "204":
          description: ""
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Body'
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - ApiKeyAuth: []
      summary: Получение случая для оценки экспертом
      tags:
      - expert
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"